apps/worker/requirements.txt
sentence-transformers
opensearch-py
tqdm
pinecone-client
python-dotenv
fastapi
uvicorn[standard]

